requirements.txt
aiogram>=3.0.0
ryanair-py>=3.0.0
apscheduler>=3.10.0
